abc246

+++ tool.cpp
@@ -104,7 +104,7 @@ int main() {
     // 配列の最大値、和
     vl vec;
     ll max_range = *max_element(all(vec));
-    ll sum = accumulate(all(vec), 0);
+    ll sum = accumulate(all(vec), ll(0));
 
     // 二分探索
     vl vec;
